Exception handling

+++ src/main/java/com/joaovictor/commerce/services/ProductService.java
@@ -3,14 +3,20 @@ package com.joaovictor.commerce.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.joaovictor.commerce.dto.ProductDTO;
 import com.joaovictor.commerce.entities.Product;
 import com.joaovictor.commerce.repositories.ProductRepository;
+import com.joaovictor.commerce.services.exceptions.DatabaseException;
+import com.joaovictor.commerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -22,7 +28,8 @@ public class ProductService {
 	@Transactional(readOnly = true) // Indicar que é um método apenas leitura.
 	public ProductDTO findById(Long id) {
 		Optional<Product> result = repository.findById(id);  // findById retorna um tipo "Optional".
-		Product product = result.get();
+		Product product = result.orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado") );
 		ProductDTO dto = new ProductDTO(product);
 		return dto;
 	}
@@ -43,15 +50,29 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id); // Objeto monitorado pela JPA
-		copyDtoEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+		try {
+			Product entity = repository.getReferenceById(id); // Objeto monitorado pela JPA
+			copyDtoEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		} catch (EntityNotFoundException e) {
+			 throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		
+		
 	}
 
-	@Transactional 
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Operação inválida: esse dado está relacionado com outro.");
+		}
+		
 	}
 	
 	private void copyDtoEntity(ProductDTO dto, Product entity) {
